Cleanup types.
